requirements.txt
coreapi==1.32.0
cronex==0.1.0
Django==1.10
django-allauth==0.27.0
django-cors-headers==1.2.2
django-crispy-forms==1.6.0
django-filter==0.14.0
django-guardian==1.4.5
django-rest-auth==0.8.1
djangorestframework==3.4.5
itypes==1.1.0
Markdown==2.6.6
oauthlib==2.0.0
Pillow==3.3.1
python-openid==2.2.5
requests==2.11.1
requests-oauthlib==0.7.0
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
